Alterando o campo de IntegerField para IntegerBigField (para ampliar as opções)

+++ apps/shelter/migrations/0002_shelter_image_shelter_phone_alter_shelter_cep_and_more.py
@@ -19,7 +19,7 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='shelter',
             name='phone',
-            field=models.PositiveIntegerField(default=1),
+            field=models.PositiveBigIntegerField(default=1),
         ),
         migrations.AlterField(
             model_name='shelter',
