+++ apps/shelter/models/shelter.py
@@ -8,7 +8,7 @@ class Shelter(models.Model):
     CEP = models.CharField(max_length=9, blank=True)
     city = models.CharField(max_length=70)
     state = models.CharField(max_length=70)
-    phone = models.PositiveIntegerField(default=1)
+    phone = models.PositiveBigIntegerField(default=1)
     image = models.ImageField(upload_to='shelters/%Y/%m/%d/', blank=True)
 
     def __str__(self):
